% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageByForester.R
\docType{methods}
\name{ageByForester}
\alias{ageByForester}
\title{Site age by experienced forester}
\usage{
ageByForester(projectID, sampleNumber, sampleTypeCode)
}
\arguments{
\item{projectID}{character,  Project ID.}

\item{sampleNumber}{character, Sample number.}

\item{sampleTypeCode}{character, Sample type code.}
}
\value{
Age provided by experienced forester
}
\description{
This function derives the site age based on experienced forester. It is equivalent to \code{vri_age} macro in original
SAS compiler.
}
\note{
Contact Bob Krahn for details
}
\author{
Yong Luo
}
