% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyDWB.R
\docType{methods}
\name{applyDWB}
\alias{applyDWB}
\title{Apply decay, waste and breakage percentage to gross merchantable volume-VRI specific}
\usage{
applyDWB(treeMS)
}
\arguments{
\item{treeMS}{data.table, Compiled full and enhanced trees with percentage of decay, waste and breakage.}
}
\value{
A data table that contains \code{VOL_NTWB} (net volume that waste 2 wood and breakage), \code{VOL_D}
(merchantable volume after removing decay), \code{VOL_DW} (merchantable volume after removing decay
and waste) and \code{ VOL_DWB} (merchantable volume after removing decay, waste and breakage).
}
\description{
This function calculates merchantable volume after removing decay, waste and breakage in VRI compiler.
The function is part of \code{cp_vegi_2017.sas} to derive \code{tree_ms7}.
}
\author{
Yong Luo
}
