% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignChangeComponent.R
\docType{methods}
\name{assignChangeComponent}
\alias{assignChangeComponent}
\title{Assign component change for repeatedly-measured trees}
\usage{
assignChangeComponent(treelist, samples)
}
\arguments{
\item{treelist}{data.table, A tree list data table contains trees' unique tree id, i.e., site_identifier, plot, and tree_number,
trees' diameters (\code{DIAMETER}), live and dead status (\code{LV_D}),
and \code{MEASUREMENT_ANOMALY_CODE}}

\item{samples}{data.table, Table contains if a sample is measured (\code{SAMP_TYP}) in fixed area plot (\code{F}) or
variable area plot (\code{V}).}
}
\value{
one table that contains tree id and component change (column \code{COMPONENT_CHANGE}).

\itemize{
\item{\code{NA: }} {at the site first measurement, A tree was measured}
\item{\code{I: }} {ingrowth, a tree passed the size shreshold in the sites' remeasurements}
\item{\code{S: }} {suivival, a tree is live in both previous measurement and current measurement}
\item{\code{M: }} {mortality, a tree is dead in current measurement but alive in the previous measurement}
\item{\code{D: }} {dead, a tree is dead in current measurement and dead in previous measurement}
\item{\code{H: }} {harvested, tree is harvested}
\item{\code{DROP: }} {droped, a tree is dropped for some reason when it is live}
}
}
\description{
This function is to assign component change for repeatedly-measured trees based on
trees' live and dead, and visit number
}
\author{
Yong Luo
}
