% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignMeasInt.R
\docType{methods}
\name{assignMeasInt}
\alias{assignMeasInt}
\title{assign measurement intensity based on field measured attributes}
\usage{
assignMeasInt(compilationType, vic, vii, vid)
}
\arguments{
\item{compilationType}{character, either \code{PSP} or \code{nonPSP}. If it is \code{PSP}, it
is consistent with original PSP compiler, otherwise, it
is consistent with VRI compiler.}

\item{vic}{data.table, vic.}

\item{vii}{data.table, vii.}

\item{vid}{data.table, vid.}
}
\value{
Two data tables: fullDimTrees contains all the trees that have both DBH and HT,
nonHTTrees contains the trees that have only DBH infor.
}
\description{
This function it to assign measurement intensity based on field measured attributes.
}
\author{
Yong Luo
}
