% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boredAgeCalculator.R
\docType{methods}
\name{boredAgeCalculator_Bore}
\alias{boredAgeCalculator_Bore}
\alias{boredAgeCalculator_Bore,numeric,numeric-method}
\alias{boredAgeCalculator_Bore,numeric,missing-method}
\alias{boredAgeCalculator_Bore,missing,numeric-method}
\title{Derive bored age using office and field bored age}
\usage{
boredAgeCalculator_Bore(officeBoredAge, fieldBoredAge)

\S4method{boredAgeCalculator_Bore}{numeric,numeric}(officeBoredAge, fieldBoredAge)

\S4method{boredAgeCalculator_Bore}{numeric,missing}(officeBoredAge)

\S4method{boredAgeCalculator_Bore}{missing,numeric}(fieldBoredAge)
}
\arguments{
\item{officeBoredAge}{numeric, Office bored age, which is measured in lab by professionals.}

\item{fieldBoredAge}{numeric, Field bored age, estimated in field by field crew.}
}
\value{
bored age
}
\description{
This function is to derive bore age based on either office bored age (\code{officeBoredAge})
or field bored age (\code{fieldBoredAge}). When both bore age information are available, the function takes
\code{officeBoredAge} as priority. The function is one of the four functions that derive bored age using
different method. The rests are \code{\link{boredAgeCalculator_Total}}, \code{\link{boredAgeCalculator_Phys}} and
\code{\link{boredAgeCalculator_Prorated}}.
}
\seealso{
\code{\link{boredAgeCalculator_Total}}
\code{\link{boredAgeCalculator_Phys}}
\code{\link{boredAgeCalculator_Prorated}}
}
\author{
Yong Luo
}
