% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boredAgeCalculator.R
\docType{methods}
\name{boredAgeCalculator_Prorated}
\alias{boredAgeCalculator_Prorated}
\alias{boredAgeCalculator_Prorated,numeric,numeric,numeric,numeric-method}
\alias{boredAgeCalculator_Prorated,numeric,numeric,numeric,missing-method}
\title{Derive bored age using pro-rated age}
\usage{
boredAgeCalculator_Prorated(ringLength_prorated, ringCount_prorated,
  boreDiameter, barkThickness)


  \S4method{boredAgeCalculator_Prorated}{numeric,numeric,numeric,numeric}(ringLength_prorated,
  ringCount_prorated, boreDiameter, barkThickness)


  \S4method{boredAgeCalculator_Prorated}{numeric,numeric,numeric,missing}(ringLength_prorated,
  ringCount_prorated, boreDiameter)
}
\arguments{
\item{ringLength_prorated}{numeric, Pro-rated ring length in cm}

\item{ringCount_prorated}{numeric, Pro-rated ring count}

\item{boreDiameter}{numeric, Diameter at bore in cm}

\item{barkThickness}{numeric, Bark thickness in mm. If missing, 0.05 is used in the function.}
}
\value{
bored age
}
\description{
This function is to derive bore age based on diameter at bore (\code{boreDiameter}),
bark thickness (\code{barkThickness}), pro-rated ring length (\code{ringLength_prorated}) and
pro-rated ring count (\code{ringCount_prorated}).  The function is one of the four functions that derive bored age using
different method. The rests are \code{\link{boredAgeCalculator_Bore}}, \code{\link{boredAgeCalculator_Total}} and
\code{\link{boredAgeCalculator_Phys}}.
}
\seealso{
\code{\link{boredAgeCalculator_Total}}
\code{\link{boredAgeCalculator_Phys}}
\code{\link{boredAgeCalculator_Prorated}}
}
\author{
Yong Luo
}
