% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkMaps.R
\docType{methods}
\name{checkMaps}
\alias{checkMaps}
\title{Load maps from map source}
\usage{
checkMaps(mapPath)
}
\arguments{
\item{mapPath}{character, The path to save all the maps.
Note that all the saved maps have time
stemps, which suggest when the files have been
modified in \code{mapSourcePath}.}
}
\value{
no value is returned from this function.
}
\description{
This function is to load maps from
\code{mapSourcePath} and save them to \code{mapPath}.
For TSA and BEC, the maps are direct from bcmaps package.
}
\note{
all the maps relies on the bcdata package.
}
\author{
Yong Luo
}
