% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compilerPathSetup.R
\docType{methods}
\name{compilerPathSetup}
\alias{compilerPathSetup}
\title{Setup an output path of the compiler}
\usage{
compilerPathSetup(compilationPath = ".")
}
\arguments{
\item{compilationPath}{character, Specifies a path to store whole compilation process.
If missing, the current work directory will be used.}
}
\value{
Four paths will be returned as following:
\itemize{

\item {raw_from_oracle} {Path to save all data that read from both oracle and txt database without merging.}

\item {compilation_sa} {Path to save merged data for key tables (i.e., vi_a to vi.g) from both oracle and txt database.}

\item {compilation_db} {Path to save compiled outputs.}

\item {compilation_archive} {Path to archive all compilation process.}
}
}
\description{
This function does two things: 1. create a folder that will store compiled data;
2. return a path that directs the compiled folder.
}
\note{
Could overwrite the existing output folder, depending on user's choise, i.e., yes or no.
}
\author{
Yong Luo
}
