% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compilerPathSetup.R
\docType{methods}
\name{compilerPathSetup_new}
\alias{compilerPathSetup_new}
\title{Setup output paths of the compiler}
\usage{
compilerPathSetup_new(
  compilationPath = ".",
  compilationDate,
  compilationType,
  recompile,
  download,
  archiveDate
)
}
\arguments{
\item{compilationPath}{character, Specifies a path to store whole compilation process.
If missing, the current work directory will be used.}

\item{compilationDate}{character, Specifies a compilation date. It should be in format of YYYYMMDD
It will be used for archive the compilation outputs.}

\item{compilationType}{character, either \code{PSP} or \code{nonPSP}. If it is \code{PSP}, it
is consistent with original PSP compiler, otherwise, it
is consistent with VRI compiler.}

\item{recompile}{logical, Defines whether we want to use existing data that downloaded
previously.}

\item{download}{logical, Defines whether we want to use existing data that downloaded
previously.}

\item{archiveDate}{character, Defines on which archive date the raw data were downloaded.
These raw data will be used for recompilation. Format is YYYYMMDD.}
}
\value{
Seven paths will be returned as following:
\itemize{

\item {raw_from_oracle} {Path to save all data that read from both oracle and txt database without merging.}

\item {compilation_sa} {Path to save merged data for key tables (i.e., vi_a to vi.g) from both oracle and txt database.}

\item {compilation_db} {Path to save compiled outputs.}

\item {compilation_archive} {Path to archive all compilation process.}

\item {compilation_report} {Path to report compilation process.}

\item {compilation_map} {Path to archive all maps for compilation process.}

\item {compilation_coeff} {Path to archive all coefficients for compilation process.}

\item {compilation_last} {Path to archive last compilation process.}
}
}
\description{
This function does two things: 1. create folders that will store compiled data;
2. return paths that directs the compiled folder.
}
\note{
Could overwrite the existing output folder, depending on user's choise, i.e., yes or no.
}
\author{
Yong Luo
}
