% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBH_Height_MEM.R
\name{DBH_Height_MEM}
\alias{DBH_Height_MEM}
\title{Develop the mixed effect models between DBH and height by strata.}
\usage{
DBH_Height_MEM(compilationPath, coeffSavePath, fityear)
}
\arguments{
\item{compilationPath}{character, Specifies the path that stores all the data/processes.}

\item{coeffSavePath}{character, Specifies the path to save your outputs.}

\item{fityear}{numeric, Specifies the year of the fit.}
}
\value{
no item returned
}
\description{
Develop the mixed effect model, and select the best models per strata.
}
\author{
Yong Luo
}
