% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBHGrowthRateGenerator.R
\docType{methods}
\name{DBHGrowthRateGenerator}
\alias{DBHGrowthRateGenerator}
\title{Derive DBH growth rate by DBH class, growth length class, species and bec zone}
\usage{
DBHGrowthRateGenerator(treeData, minObs = 100)
}
\arguments{
\item{minObs}{numeric, Specify minimum number of observation for calculating mean growth rate.
If missing, default is 100.}

\item{treesData}{data.table, Contains both height trees and nonHT trees in PSPs.}
}
\value{
Growth rate lookup table
}
\description{
Derive DBH growth rate by DBH class, growth length class, species and bec zone
}
\author{
Yong Luo
}
