% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWBCompiler.R
\docType{methods}
\name{DWBCompiler}
\alias{DWBCompiler}
\alias{DWBCompiler,data.table,data.table,data.table,character-method}
\alias{DWBCompiler,data.table,data.table,data.table,missing-method}
\title{Compile decay, waste and breakage for standard tables-VRI specific}
\usage{
DWBCompiler(treeMS, siteAge, treeLossFactors, equation)

\S4method{DWBCompiler}{data.table,data.table,data.table,character}(treeMS, siteAge, treeLossFactors, equation)

\S4method{DWBCompiler}{data.table,data.table,data.table,missing}(treeMS, siteAge, treeLossFactors)
}
\arguments{
\item{treeMS}{data.table, Tree-level data that has been compiled whole stem volume and gross merchantable volume for full and enhanced trees.}

\item{siteAge}{data.table, Cluster-level summaries of age and height. This table is an output from
\code{\link{siteAgeSummary}}}

\item{treeLossFactors}{data.table, The tree loss factor data, an output of \code{\link{VRIInit_lossFactor}}.
In this funtion, this table provides loss indicator.}

\item{equation}{character, Specifies whether the compiler is based on \code{KFIZ} or \code{KBEC}.
Default is set as \code{KBEC}.}
}
\value{
A compiled volume after removing decay, waste and breakage; a log file
}
\description{
This function compiles decay, waste and breakage for standard tables in VRI compiler. The function
is equivalent to \code{dwb_vri_2017.sas}.
}
\author{
Yong Luo
}
