% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWBGenerator_BEC.R
\docType{methods}
\name{DWBGenerator_BEC}
\alias{DWBGenerator_BEC}
\alias{DWBGenerator_BEC,numeric,numeric,character,numeric,character,character,character-method}
\title{Collect decay, waste and breakage factor in BEC routine}
\usage{
DWBGenerator_BEC(DBH, height, species, meanAge, BEC, riskGroup, adjustID)

\S4method{DWBGenerator_BEC}{numeric,numeric,character,numeric,character,character,character}(DBH, height, species, meanAge, BEC, riskGroup, adjustID)
}
\arguments{
\item{DBH}{numeric, Tree DBH.}

\item{height}{numeric, Tree height.}

\item{species}{character, Tree basic species code, which is SP0 in VRI original data.}

\item{meanAge}{numeric, Mean site age.}

\item{BEC}{character, BC BEC zone.}

\item{riskGroup}{character, Specifies the risk group.  It must be one of \code{1}, \code{2} or \code{3}.
It can be derived from \code{\link{riskGroupDeriver}}.}

\item{adjustID}{character, Adjustment identifier. Blank is no adjustment; QCI is queen charlottes; WET is wetbelt
and GLD_NW golden ?.}
}
\value{
A list of \code{decay}, \code{waste} and \code{breakage} percentage.
}
\description{
This function is to collect the dead, waste and breakage factor from lookup table and
join them into tree data. Instead of reading the lookup table from disk, the function
uses hard-coded the lookup table. This function is equivalent to \code{dwb_v3.sas} macro.
For \code{FIZ} routine, the decay, waste and breakage are collected using funtion
\code{\link{DWBGenerator_FIZ}}
}
\author{
Yong Luo
}
