% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDWBSeries.R
\docType{methods}
\name{getDWBSeries}
\alias{getDWBSeries}
\alias{getDWBSeries,character,character,character,character,character-method}
\alias{getDWBSeries,character,character,character,missing,character-method}
\alias{getDWBSeries,character,character,missing,character,character-method}
\title{Get local DWB series}
\usage{
getDWBSeries(species, ageRangeClass, PSYUB, FIZ, source)

\S4method{getDWBSeries}{character,character,character,character,character}(species, ageRangeClass, PSYUB, FIZ, source)

\S4method{getDWBSeries}{character,character,character,missing,character}(species, ageRangeClass, PSYUB, source)

\S4method{getDWBSeries}{character,character,missing,character,character}(species, ageRangeClass, FIZ, source)
}
\arguments{
\item{species}{character, Tree basic species code, which is SP0 in VRI original data.}

\item{ageRangeClass}{character, The classified age range into 1 to 4. Output from \code{\link{ageRangeClassifier}}.}

\item{PSYUB}{character, Unique PSYU+PSYU_BLK code.}

\item{FIZ}{character, BC forest inventory zone.}

\item{source}{character, Series source, must one of \code{local}, \code{zonal} and \code{reversingZonal}.
\code{local} is based on species, PSYUB and ageRangeClass;
\code{zonal} is based on FIZ, species and ageRangeClass;
and \code{reversingZonal} is based on reversing FIZ zones, i.e.,
changing interior to coastal and changing coastal to interior.}
}
\value{
DWB series, a two number character.
}
\description{
Join local DWB (old/imperical factors) by \code{species}, \code{PSYUB}, \code{ageRangeClass} and \code{FIZ}.
This function is the first part of \code{dwb_fct.sas}.
}
\author{
Yong Luo
}
