% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grossVolCal_kozak.R
\docType{methods}
\name{grossVolCal_kozak}
\alias{grossVolCal_kozak}
\title{Calcualte gross volume using kozak equations}
\usage{
grossVolCal_kozak(
  compilationType,
  fullDimTreeData,
  logMinLength,
  stumpHeight,
  breastHeight,
  UTOPDIB
)
}
\arguments{
\item{compilationType}{character, either \code{PSP} or \code{nonPSP}. If it is \code{PSP}, it
is consistent with original PSP compiler, otherwise, it
is consistent with VRI compiler.}

\item{fullDimTreeData}{data.table, This is full dimension tree data, i.e., it has both DBH and height.}

\item{logMinLength}{numeric, Specifies a minimum length for a log.}

\item{stumpHeight}{numeric, Specifies stump height. If missing, 0.3 m will be used.}

\item{breastHeight}{numeric, Specifies breast height. 1.3 m will be used when this arguement is missing.}

\item{UTOPDIB}{numeric, Specifies minimum merchantable inside bark diameter. 10 cm is used as a default.}
}
\value{
A data table
}
\description{
This function use kozak BEC and species-specific taper equation to
calculate tree volume.
This functions also assigns the volume multiplier adjustment.
}
\author{
Yong Luo
}
