% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GYS_oracle_org1.R
\docType{methods}
\name{GYS_oracle_org1}
\alias{GYS_oracle_org1}
\title{title GYS_oracle_org1}
\usage{
GYS_oracle_org1(oracleSourcePath, outputPath)
}
\arguments{
\item{oracleSourcePath}{character, Specifies the path that stores data from oracle data base.
In VRI compiler, this should be the savePath for \code{\link{loadVGIS}}.}

\item{outputPath}{character, Specifies the path to save your outputs. If missing, the current working
directory will be choosed.}
}
\value{
no item returned
}
\description{
This function is to merge oracle and ascii data before the GYS compiler, the function below is modified
from ld_gysfx_pt1_rene_10jun2017.sas
The major changes from the original codes are
1. removed spatial attributes in GYS and updated from the most recent map
2. clearified the processes in sample_id, plot and tree_no order, as the original sas version creates
massive tables, and confuses the compilation.
3. removed some pre-compiled results as inputs
}
