% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heightEstimate_byHeightModel.R
\docType{methods}
\name{heightEstimate_byHeightModel}
\alias{heightEstimate_byHeightModel}
\title{Estimate height using height-DBH curves}
\usage{
heightEstimate_byHeightModel(beczone, subzone, species, DBH, heightModels)
}
\arguments{
\item{beczone}{character, BEC zone.}

\item{subzone}{character, BEC subzone.}

\item{species}{character, Species.}

\item{DBH}{numeric, Diameter at breast height.}

\item{heightModels}{data.table, Specifies the best model and coefficients by each
BEC subzone and species.}
}
\value{
projected total height
}
\description{
Estimate height using height-DBH curves by bec subzone and species
}
\author{
Yong Luo
}
