% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heightEstimate_byHeightModel.R
\docType{methods}
\name{heightEstimate_mixedEffect}
\alias{heightEstimate_mixedEffect}
\title{Estimate height using mixed effect height-DBH curves}
\usage{
heightEstimate_mixedEffect(BEC, siteID, sp0, DBH, fixedEffects, randomEffects)
}
\arguments{
\item{BEC}{character, BEC zone.}

\item{siteID}{character, site identifier}

\item{DBH}{numeric, Diameter at breast height (cm).}

\item{fixedEffects}{data.table, Fixed effect of the best model by each
BEC and sp0.}

\item{randomEffects}{data.table, Random effects of the best model by each
BEC, sp0 and site identifier.}

\item{species}{character, Species.}
}
\value{
projected total height
}
\description{
Estimate height using height-DBH curves by bec and species
}
\author{
Yong Luo
}
