% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heightEstimate_byHeightModel.R
\docType{methods}
\name{heightEstimate_mixedEffect_nlme}
\alias{heightEstimate_mixedEffect_nlme}
\title{Estimate height using mixed effect height-DBH curves using nlme}
\usage{
heightEstimate_mixedEffect_nlme(
  siteID,
  unitreeid,
  species,
  DBH,
  fixedEffects,
  randomEffects_site,
  randomEffects_tree
)
}
\arguments{
\item{siteID}{character, site identifier}

\item{species}{character, Species.}

\item{DBH}{numeric, Diameter at breast height (cm).}

\item{fixedEffects}{data.table, Fixed effect of the best model by each
sp0.}

\item{randomEffects_site}{data.table, Random effects of the best model at site level.}

\item{randomEffects_tree}{data.table, Random effects of the best model at tree level.}
}
\value{
projected total height
}
\description{
Estimate height using height-DBH curves by species
}
\author{
Yong Luo
}
