% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heightSmry_byC.R
\docType{methods}
\name{heightSmry_byC}
\alias{heightSmry_byC}
\alias{heightSmry_byC,data.table-method}
\title{Summarize mean and lorry's height by cluster-VRI specific}
\usage{
heightSmry_byC(treeMC)

\S4method{heightSmry_byC}{data.table}(treeMC)
}
\arguments{
\item{treeMC}{data.table, Compiled tree-level data that contains both measured trees and counted trees.}
}
\value{
A table contains computed mean height \code{MN or MEAN} and lorey's height \code{LRY} for all standing trees \code{ALL},
standing and live trees \code{1} and standing and non-broken top trees \code{2}. The output is equevalent
to \code{height} table in original compiler.
}
\description{
Summarizes mean and lorry's height by cluster for standing trees, standing + live trees, and standing
+ live + non-broken top trees.
The function is improved version to calculate mean height in \code{vol_ha_2017.sas} by outputing lorey's
height.
For both fixed and variable area plots, the function computes mean height by using plot weight (\code{PLOT_WT})
weighted height. For lorey's height computation, the function treats variable and fixed area plots differently.
Specifically, the function uses the mean height as lorey's height for variable plots,
while uses height that weighted both by plot weight (\code{PLOT_WT}) and basal area (\code{BA_TREE})
for fixed area plots.
}
\author{
Yong Luo
}
