% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISMC_VGISTranslator.R
\name{ISMC_VGISTranslator}
\alias{ISMC_VGISTranslator}
\title{Translate ismc data formats into vgis formats}
\usage{
ISMC_VGISTranslator(inputPath, outputPath)
}
\arguments{
\item{inputPath}{character, Specifies the path that stores data from oracle data base.}

\item{outputPath}{character, Specifies the path to save your outputs. If missing, the current working
directory will be choosed.}
}
\value{
no item returned
}
\description{
This function is to translate ismc data formats into vgis formats
}
\author{
Yong Luo
}
