% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISMCCompiler_new.R
\docType{methods}
\name{ISMCCompiler_new}
\alias{ISMCCompiler_new}
\title{ISMC compiler}
\usage{
ISMCCompiler_new(
  compilationType,
  oracleUserName,
  oraclePassword,
  oracleEnv = "INT",
 
    compilationPath = "//albers/gis_tib/VRI/RDW/RDW_Data2/Work_Areas/VRI_ASCII_PROD/FromRCompiler",
  equation = "KBEC",
  walkThru = TRUE,
  logMinLength = 0.1,
  stumpHeight = 0.3,
  breastHeight = 1.3,
  UTOPDIB = 10,
  utilLevel = 4,
  weirdUtil = "4",
  recompile = FALSE,
  archiveDate = as.character(NA)
)
}
\arguments{
\item{compilationType}{character, either \code{PSP} or \code{nonPSP}. If it is \code{PSP}, it
is consistent with original PSP compiler, otherwise, it
is consistent with VRI compiler.}

\item{oracleUserName}{character, User name to access to ISMC database.}

\item{oraclePassword}{character, Password to access to ISMC database.}

\item{oracleEnv}{character, Specify which environment of ISMC database the data download from.
Currently, it supports 1) \code{INT} for intergration environment;
2) \code{TST} for test environment; 3) \code{PROD} for final production
environment.}

\item{compilationPath}{character, Specifies the path that stores all the data/processes. By specifying this,
four folders will be created to record all the data/processes. Specifically,
raw_from_oracle stores the data just after oracle and ascii without editing;
compilation_sa stores key data (not all) that after editing and before volume and age compilation;
compilation_db stores compiled results for volume and age compilation at both tree level
and cluater level;
Archive_YYYYMMDD achives all the data mentioned above for the future use or reference.
By default, this path is set as \code{//albers/gis_tib/VRI/RDW/RDW_Data2/Work_Areas/VRI_ASCII_PROD/RCompilation},
which is consistent with our rdw system.}

\item{equation}{character, Specifies the taper equation that is used for compiler. Currently supports
BEC-based (\code{KBEC}) and FIZ-based (\code{KFIZ}).}

\item{walkThru}{logical, Speciefies whether the data had been collected using work through method. Default is \code{TRUE},
if it is not specified.}

\item{logMinLength}{numeric, Specifies minimum length of log when doing log length adjustment,
see \code{\link{logMatrixAdjustment}} for details. If missing 0.1 is used.}

\item{stumpHeight}{numeric, Stump height. If missing 0.3 is used.}

\item{breastHeight}{numeric, Breast height. If missing 1.3 is used.}

\item{UTOPDIB}{numeric, Threshold inside-bark diameter for merchantable volume. If missing, UTOPDIB is 10.}

\item{utilLevel}{numeric, Specifies utilization level in summrizing tree volumes at cluster and species level. Default is 4.}

\item{weirdUtil}{character, Specifies weird utilization in summarizing tree volumes at cluster and species level.
Default is \code{no}, if missing. Otherwise, a number should be provided.}

\item{recompile}{logical, Defines whether we want to recompile data using archived
raw data. Default is FALSE, which means the compiler needs to
download data from ISMC database. When it is \code{TRUE}, a folder will
be created in format of Archive_YYYYMMDD(archiveDate)_RecompYYYYMMDD(current date) to save the all the compilation.}

\item{archiveDate}{character, Defines on which archive date the raw data were downloaded.
These raw data will be used for recompilation. Format is YYYYMMDD.}
}
\value{
This function compiles data and save outputs in \code{compilationPaths$compilation_db} and no file is returned.
}
\description{
This compiler is a general compiler to compile field data from either PSP or other
programs.
}
\note{
Improvements include:
\enumerate{
\item specifies trees in B plots as height enhanced trees
\item removes the sas-dependent sindex functions
\item introduce the SIndexR package
}
Currently, the compiler supports compilation for the below sample types:
\itemize{
\item{\code{Q: }} {Regular VRI sample with five point clusters design}
\item{\code{T: }} {??, sample protocol and sample design are same as Q samples}
\item{\code{B: }} {??, same plot layout as Q sample, with height is measured for all trees and no call grading information}
\item{\code{M: }} {CMI sample, with all trees have call grading information in the field}
\item{\code{L: }} {LiDAR project, same plot layout and same design but without call grading information}
\item{\code{Y: }} {YSM plots, population between 15 and 50 years}
\item{\code{F: }} {NFI plots, trees measured all DBH, height and call grading}
\item{\code{N: }} {NVAF plots}
\item{\code{A: }} {VRI audit plots}
}
}
\references{
VRI compiler manual
}
\author{
Yong Luo
}
