% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISMCDataPrep.R
\docType{methods}
\name{ISMCDataPrep}
\alias{ISMCDataPrep}
\title{prepare raw data from ismc for ISMCCompiler}
\usage{
ISMCDataPrep(compilationType, inputPath, outputPath, coeffPath)
}
\arguments{
\item{compilationType}{character, either \code{PSP} or \code{nonPSP}. If it is \code{PSP}, it
is consistent with original PSP compiler, otherwise, it
is consistent with VRI compiler.}

\item{inputPath}{character, Specifies the path that stores data from oracle data base.}

\item{outputPath}{character, Specifies the path to save your outputs. If missing, the current working
directory will be choose.}

\item{coeffPath}{character, Specifies the path where the crosswalk
table is stored.}
}
\value{
no item returned
}
\description{
This function is to prepare ismc data for compilation and save them
in \code{outputPath} folder
}
\author{
Yong Luo
}
