% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logAdjustment.R
\docType{methods}
\name{logAdjustment}
\alias{logAdjustment}
\alias{logAdjustment,data.table,numeric-method}
\alias{logAdjustment,data.table,missing-method}
\title{Adjust log length - VRI specific}
\usage{
logAdjustment(treeData, stumpHeight)

\S4method{logAdjustment}{data.table,numeric}(treeData, stumpHeight)

\S4method{logAdjustment}{data.table,missing}(treeData)
}
\arguments{
\item{treeData}{data.table, Must have tree data information. The table is an output of \code{\link{VRIInit_measuredTree}}.}

\item{stumpHeight}{numeric, Length of stump. As default, this arguement is set as 0.3 m.}
}
\value{
Data table that contains the adjusted log length
}
\description{
The function is to adjust the log length to fit actual height.
This function is equivalent to \code{log_adj_new} macro in original VRI compiler
}
\note{
Please see Bob for details about input files
}
\author{
Yong Luo
}
