% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logAttributesLongForm.R
\docType{methods}
\name{logAttributesLongForm}
\alias{logAttributesLongForm}
\alias{logAttributesLongForm,data.table,numeric-method}
\alias{logAttributesLongForm,data.table,missing-method}
\title{Transpose wide form table to long form table-VRI specific}
\usage{
logAttributesLongForm(treeData, maximumLogNO)

\S4method{logAttributesLongForm}{data.table,numeric}(treeData, maximumLogNO)

\S4method{logAttributesLongForm}{data.table,missing}(treeData, maximumLogNO)
}
\arguments{
\item{treeData}{data.table, an output from \code{logValueCalculator} function, i.e., tree_ms6. This table currently
has top diameter (LOG_D_X), length (LOG_L_X), volume (LOG_V_X), merchantable volume (LOG_VM_X),
grade (LOG_G_X), sound percentage (LOG_S_X) and value (LOG_c_x). X is log number from 1 to maximum log number.}

\item{maximumLogNO}{numeric, determine the maximum number of logs. In VRI compiler, it is 9. Therefore, 9 is default.}
}
\value{
A data table and a log file
}
\description{
This function transposes wide form outputs to long form outputs. This function is not
included in the VRI compiler anymore.
}
\author{
Yong Luo
}
