% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logFileProducer.R
\docType{methods}
\name{logFileProducer}
\alias{logFileProducer}
\alias{logFileProducer,character,character,data.table,character,character,character-method}
\alias{logFileProducer,character,character,data.table,missing,character,character-method}
\alias{logFileProducer,character,character,data.table,character,missing,missing-method}
\alias{logFileProducer,character,character,data.table,missing,missing,missing-method}
\title{To display a table to log file}
\usage{
logFileProducer(reason, action, displayTable, displayColumn,
  changedVariable, fromTo)


  \S4method{logFileProducer}{character,character,data.table,character,character,character}(reason,
  action, displayTable, displayColumn, changedVariable, fromTo)


  \S4method{logFileProducer}{character,character,data.table,missing,character,character}(reason,
  action, displayTable, changedVariable, fromTo)


  \S4method{logFileProducer}{character,character,data.table,character,missing,missing}(reason,
  action, displayTable, displayColumn)


  \S4method{logFileProducer}{character,character,data.table,missing,missing,missing}(reason,
  action, displayTable)
}
\arguments{
\item{reason}{character, Reason to trigger an action.}

\item{action}{character, Specifies action from one of \code{removed}, \code{no} and \code{changed}.}

\item{displayTable}{data.table, A table of interest}

\item{displayColumn}{character, Specifies which column(s) will be displayed in the log file.}

\item{changedVariable, }{charcater, Specifies the variable that has been modified, must be present if action
is \code{changed}.}

\item{fromTo}{character, This is two vectors character. Specifies columns that before and after alteration.
Must be present when action is set as \code{changed}.}
}
\value{
A tring of text
}
\description{
Simple tool to display a table to log file.
}
\author{
Yong Luo
}
