% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logMatrixAdjustment.R
\docType{methods}
\name{logMatrixAdjustment}
\alias{logMatrixAdjustment}
\alias{logMatrixAdjustment,data.table,numeric,numeric,numeric,numeric-method}
\alias{logMatrixAdjustment,data.table,numeric,missing,numeric,numeric-method}
\alias{logMatrixAdjustment,data.table,numeric,numeric,missing,numeric-method}
\alias{logMatrixAdjustment,data.table,numeric,numeric,numeric,missing-method}
\alias{logMatrixAdjustment,data.table,numeric,missing,missing,numeric-method}
\alias{logMatrixAdjustment,data.table,numeric,missing,numeric,missing-method}
\alias{logMatrixAdjustment,data.table,numeric,missing,missing,missing-method}
\title{Adjust log length matrix}
\usage{
logMatrixAdjustment(
  logLengthMatrix,
  height,
  stumpHeight,
  logMinLength,
  logDefaultLength
)

\S4method{logMatrixAdjustment}{data.table,numeric,numeric,numeric,numeric}(
  logLengthMatrix,
  height,
  stumpHeight,
  logMinLength,
  logDefaultLength
)

\S4method{logMatrixAdjustment}{data.table,numeric,missing,numeric,numeric}(logLengthMatrix, height, logMinLength, logDefaultLength)

\S4method{logMatrixAdjustment}{data.table,numeric,numeric,missing,numeric}(logLengthMatrix, height, stumpHeight, logDefaultLength)

\S4method{logMatrixAdjustment}{data.table,numeric,numeric,numeric,missing}(logLengthMatrix, height, stumpHeight, logMinLength)

\S4method{logMatrixAdjustment}{data.table,numeric,missing,missing,numeric}(logLengthMatrix, height, logDefaultLength)

\S4method{logMatrixAdjustment}{data.table,numeric,missing,numeric,missing}(logLengthMatrix, height, logMinLength)

\S4method{logMatrixAdjustment}{data.table,numeric,missing,missing,missing}(logLengthMatrix, height)
}
\arguments{
\item{logLengthMatrix}{data.table, A matrix of log length for each tree, NA is accepted in the matrix.
The order of log from bottom to top must be presented from left to right in the table}

\item{height}{numeric, Tree height}

\item{stumpHeight, }{numeric, Stump height. If missing, 0.3 m is used.}

\item{logMinLength}{numeric, Minimum log length. If missing, 3 m is used.}

\item{logDefaultLength}{numeric, Default log length. If missing, 5 m is used.}
}
\value{
A data.table that contains the matrix of adjusted log length
}
\description{
This function is to adjust log length matrix based on tree height, minimum log length and default log length.
This function is equivalent to vol_tree_log_validation macro in original sas compiler.
}
\author{
Yong Luo
}
