% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookupCheck.R
\docType{methods}
\name{lookupCheck}
\alias{lookupCheck}
\alias{lookupCheck,character,character-method}
\alias{lookupCheck,character,missing-method}
\title{Test whether the lookup table is updated}
\usage{
lookupCheck(lookupName, lookupPath)

\S4method{lookupCheck}{character,character}(lookupName, lookupPath)

\S4method{lookupCheck}{character,missing}(lookupName)
}
\arguments{
\item{lookupName}{character, Specifies the name of lookup table.}

\item{lookupPath}{character, Path that directs to lookup tables.}
}
\value{
Not value returned. A warning message is given if the lookup table is changed.
}
\description{
Reports whether a lookup table is updated. This function is highly recommended before
proceed the VRI compiler.
}
\author{
Yong Luo
}
