% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measNumCorrect_PSP.R
\docType{methods}
\name{measNumCorrect_PSP}
\alias{measNumCorrect_PSP}
\title{Measurement number correction for raw PSP data}
\usage{
measNumCorrect_PSP(sampleID, measureNumber)
}
\arguments{
\item{sampleID}{character, Specifies how the PSP sample ID.}

\item{measureNumber}{numeric, Specifies measurement number before correction.}
}
\value{
corrected measurement number
}
\description{
This function is to correct the measurement number for the remeasured PSP data.
The remeasured PSP data is defined as PSP that sits outside of GYS oracle database.
In those remeasurement data, some measurement number has not been correctly assigned, and
conficts the measurement number for the same samples in GYS oracle database.
}
\note{
with confirmation by Rene
}
\author{
Yong Luo
}
