% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeAllVolTrees.R
\docType{methods}
\name{mergeAllVolTrees}
\alias{mergeAllVolTrees}
\alias{mergeAllVolTrees,data.table,data.table-method}
\title{Merge all volume trees-VRI specific}
\usage{
mergeAllVolTrees(treeMS, treeAX)

\S4method{mergeAllVolTrees}{data.table,data.table}(treeMS, treeAX)
}
\arguments{
\item{treeMS}{data.table, Compiled full, enhanced and H-enhanced trees. This data should be listed in vi_c table.
This data is an output of \code{\link{DWBCompiler}}}

\item{treeAX}{data.table, Non-enhanced trees in anxilirary plots (vi_i). Supposedly, the table only contains
non-enhanced tree list. However, some enhanced and H-enhanced trees also been
stored in this dataset.
An output from \code{\link{VRIInit_auxTree}}.}
}
\value{
A data table that contains all volume trees without duplicates. Equivalent to \code{tree_vb} table.
}
\description{
Merge all the volume trees, including full trees (fully-measure trees in IPC),
enhanced trees (fully-measured trees in auxi plots), H-enhanced trees (Height measured in
auxi plots) and non-enhanced trees (only DBH measured in auxi plots).
The function is part of \code{vol_ha_2017.sas} and modified dramatically in R compiler.
}
\author{
Yong Luo
}
