% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeOAData.R
\name{mergeOAData}
\alias{mergeOAData}
\title{Merge oracle and ascii data for vi_(a, ..i)}
\usage{
mergeOAData(
  oracleSourcePath,
  asciiSourcePath,
  coeffPath,
  fizmapPath,
  fizmapName,
  fizmapFormat,
  outputPath
)
}
\arguments{
\item{oracleSourcePath}{character, Specifies the path that stores data from oracle data base.
In VRI compiler, this should be the savePath for \code{\link{loadVGIS}}.}

\item{asciiSourcePath}{character, Specifies the path that stores data from ascii data base.
In VRI compiler, this should be the savePath for \code{\link{loadASCII}}.}

\item{coeffPath}{character, Specifies the path that stores coefficients and spatial lookup tables.}

\item{fizmapPath}{character, Specifies the path to forest inventory zone map. By default,
it is set to \code{//spatialfiles2.bcgov/work/for/vic/hts/dam/workarea/data/infrastructure},
which is maintained by FAIB employee.}

\item{fizmapName}{character, Specifies the name of forest inventory zone map. By default,
it is set to \code{FIZ_REG_COMPARTMENT}, which is maintained by FAIB employee.}

\item{fizmapFormat}{character, Specifies the format of forest inventory zone map. Currently, it can be specified
as \code{gdb} for geodatabase format and \code{shp} for shapefile format.
By default, it is set to \code{gdb}, which is maintained by FAIB employee.}

\item{outputPath}{character, Specifies the path to save your outputs. If missing, the current working
directory will be choosed.}
}
\value{
no item returned
}
\description{
This function is to merge oracle and ascii data before the VRI compiler
}
\author{
Yong Luo
}
