% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netVolumeCalculator.R
\docType{methods}
\name{netVolumeCalculator}
\alias{netVolumeCalculator}
\alias{netVolumeCalculator,data.table,data.table,data.table-method}
\alias{netVolumeCalculator,missing,data.table,data.table-method}
\alias{netVolumeCalculator,data.table,missing,data.table-method}
\alias{netVolumeCalculator,data.table,data.table,missing-method}
\title{Calcualtes total net volume and merchantable volume-VRI specific}
\usage{
netVolumeCalculator(grossVolMatrix, grossMerchVolMatrix, netFactorMatrix)


  \S4method{netVolumeCalculator}{data.table,data.table,data.table}(grossVolMatrix,
  grossMerchVolMatrix, netFactorMatrix)


  \S4method{netVolumeCalculator}{missing,data.table,data.table}(grossMerchVolMatrix,
  netFactorMatrix)


  \S4method{netVolumeCalculator}{data.table,missing,data.table}(grossVolMatrix,
  netFactorMatrix)


  \S4method{netVolumeCalculator}{data.table,data.table,missing}(grossVolMatrix,
  grossMerchVolMatrix)
}
\arguments{
\item{grossVolMatrix}{data.table, Calculated gross volume for each log. The first column of this table is
the volume for the stump. If missing, the function calculates the total
net merchantable volume.}

\item{grossMerchVolMatrix}{data.table, Calculated gross merchantable volume for each log. If missing, all the merchantable
volume is assigned as 0.}

\item{netFactorMatrix}{data.table, Ground call for sound percentage. If missing, the net factoring will be assigned as 100.}
}
\value{
Data table that contains total net volume (\code{VOL_NET}) and total net merchantable volume (\code{VOL_NETM}) for each tree.
}
\description{
This function calculates total net volume and net merchantable volume for each tree
based on ground called sound percentage. From the second column to the last column,
the \code{grossVolMatrix} table should have same dimensions (i.e.,
number of rows and columns) of \code{netFactorMatrix} and \code{grossMerchVolMatrix} if
they are provided. Furthermore, be aware of the correspondingness among the matrix.
This function is part of \code{log_valu_2017.sas}.
}
\author{
Yong Luo
}
