% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathIndicatorGenerator.R
\docType{methods}
\name{pathIndicatorGenerator}
\alias{pathIndicatorGenerator}
\title{Generate path indicator string for risk group}
\usage{
pathIndicatorGenerator(
  lossIndicatorMatix,
  lossIndicatorLocMatrix,
  merchantableHeight,
  compiler
)
}
\arguments{
\item{lossIndicatorLocMatrix}{data.table, Contains the location for each corresponding loss indicator
in \code{lossIndicatorMatrix}. This matrix will be used for VRI compiler.}

\item{merchantableHeight}{numeric, Specifies the maximum height for merchantable volume.
This arguement will be used for VRI compiler.}

\item{compiler}{character, Specifies in which compiler the path indicator will be generated.
It can be either \code{PSP} or \code{VRI}.}

\item{lossIndicatorMatrix}{data.table, Contains eight columns of loss indicator, i.e., LOSS1...8_IN.}
}
\value{
A length of eight character that contains 0 and 1, e.g., 10010000.
}
\description{
This function to generate a length of eight character path indicator string.
The path indicator string is an input for rish group for Jim's decay, waste and breakage function.
The function varies with PSP compiler and VRI compiler.
}
\note{
For PSP compiler, the path indicator is only based on loss indicator. However, for
VRI compiler, the path indicator also based on loss indicator location and merchantable height.
}
\author{
Yong Luo
}
