% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDynamics.R
\docType{methods}
\name{plotDynamics}
\alias{plotDynamics}
\title{Summarize the dynamic table between two measurements for repeatedly measured samples}
\usage{
plotDynamics(treelist, samples)
}
\arguments{
\item{treelist}{data.table, Tree measurements with component change, which is the
output from \code{assignChangeComponent}.}

\item{samples}{data.table, Site visit information including \code{SITE_IDENTIFIER}
and \code{VISIT_NUMBER}.}
}
\value{
return a summary table
}
\description{
This function takes tree-level component change data and summarizes the
data to a plot-level dynamics.
}
\author{
Yong Luo
}
