% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparePublishData.R
\docType{methods}
\name{preparePublishData}
\alias{preparePublishData}
\title{Prepare the compiled data for publish}
\usage{
preparePublishData(
  compilationPath,
  compilationDate,
  publishPath,
  compilationType
)
}
\arguments{
\item{compilationPath}{character, The path to the compiled data, which is configured
from \code{ISMCCompiler}.}

\item{compilationDate}{numeric, The date of the compiled data.}

\item{publishPath}{character, The path to save prepared data.}

\item{compilationType}{character, Specifies the compilation type either \code{nonPSP} or code{PSP}.}
}
\value{
no value returned. Instead, all the files will be saved into the \code{tempPath} including a readme file.
}
\description{
To prepare the compiled data for publish, currently supporting nonPSP part
}
\note{
The compilationPath must have all the outputs from \code{ISMCCompiler}.
}
\author{
Yong Luo
}
