% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareTASSInputs.R
\docType{methods}
\name{prepareTASSInputs}
\alias{prepareTASSInputs}
\title{To prepare input data for TASS run}
\usage{
prepareTASSInputs(
  inputPath,
  outputPath,
  projectName,
  clstrIDs = NA,
  siteIndexTable = NA,
  siteIndexTableSource = "ISMCCompiler",
  siteIndexMethod,
  treeVigorMethod,
  vigorAdjust08 = TRUE,
  randomSeed = NA
)
}
\arguments{
\item{inputPath}{character, The path to the compilation output.}

\item{outputPath}{character, The path to save the outputs.}

\item{projectName}{character, The name of a project, currently support \code{YSM}, \code{Taan} or \code{special}.}

\item{clstrIDs}{character, A list of clsterid in a special project. It will be ignored if the \code{projectName} is \code{YSM}
or \code{Taan}. As they are rule-based selection of clster id.}

\item{siteIndexTable}{data.table, The table to contain site index.}

\item{siteIndexTableSource}{character, Indicates where the site index table from either \code{Rene} or \code{ISMCCompiler}.
If the source is ISMCCompiler, the function will ignore the site index table provided.
Therefore, it can be missing. Default is \code{ISMCCompiler}.}

\item{siteIndexMethod}{character, The method to derive site index for repeatedly-visited sites.
There are three methods can be used: \code{byvisit}, \code{average}, \code{firstvisit}
or \code{closest50}. \code{byvisit} uses the site index for each visit.
\code{average} takes the mean site index for a given
site and species over multiple visits. \code{firstvisit} uses the site index
from the first visit. \code{closest50} uses the site index which has the closest
to the stand age of 50 years.}

\item{treeVigorMethod}{character, Method to derive a tree's vigor,
currently support \code{mainsub}.}

\item{vigorAdjust08}{logical, Indicates if the mean of mean of the tree height vigor needed to be adjusted to 0.8.
Default is \code{TRUE}.}

\item{randomSeed}{numeric, The random seeds for the stem mapping extension. Default is \code{NA}, which
does not have a seed number.}
}
\description{
This function takes the compilation outputs and prepare stand age, loss factor,
site index and stem mapping, so that the outputs are ready for TASS run
}
\note{
mainsub in treeVigorMethod:
•	For each combination of site_identifier by visit_number by species, select the 6 tallest trees from the main plot and the single tallest tree from the subplot.
•	From this subset, drop all trees with
o	BROKEN_TOP_IND = Y
o	CR_CL = I, S
o	RESIDUAL = Y
o	WALKTHRU = O
•	Compute site_height as the average height of remaining trees
•	If no trees remain, then compute an alternate site_height as the average height of all trees in the plot by site_identifier * visit_number * species (ie., no exclusion for height, broken top, crown class, residual class, or walkthru class.
}
\author{
Yong Luo
}
