% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_ID2BEC.R
\docType{methods}
\name{prj_ID2BEC}
\alias{prj_ID2BEC}
\alias{prj_ID2BEC,character-method}
\title{Assign BEC based on project ID}
\usage{
prj_ID2BEC(projectID)

\S4method{prj_ID2BEC}{character}(projectID)
}
\arguments{
\item{projectID}{character, Specifies project ID.}
}
\value{
BEC, \code{Unknown} will be return if project id does have any match in lookup table.
}
\description{
This function takes lookup table that connects project id to BEC zone and joins BEC zone
by project id. The function uses a hardcoded lookup table \code{vri_bec}.
The function is equivalent to \code{group_bec.sas}.
}
\author{
Yong Luo
}
