% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_ID2Grp.R
\docType{methods}
\name{prj_ID2Grp}
\alias{prj_ID2Grp}
\alias{prj_ID2Grp,character-method}
\title{Group project ID into project group}
\usage{
prj_ID2Grp(projectID)

\S4method{prj_ID2Grp}{character}(projectID)
}
\arguments{
\item{projectID}{character, Specifies project ID.}
}
\value{
Project group, \code{Unknown} will be return if project id does have any match in lookup table.
}
\description{
This function takes lookup table that connect project id to project group and joins project
group by project id. The function uses hardcoded lookup table \code{vri_grp}.
The function is equivalent to \code{group_prj.sas}.
}
\author{
Yong Luo
}
