% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSPCompilation_Init.R
\docType{methods}
\name{PSPCompilation_Init}
\alias{PSPCompilation_Init}
\title{Initiate PSP data in ISMC for the compilation}
\usage{
PSPCompilation_Init(inputPath, outputPath)
}
\arguments{
\item{inputPath}{character, Specifies the path that stores data from oracle data base.}

\item{outputPath}{character, Specifies the path to save your outputs. If missing, the current working
directory will be choose.}
}
\value{
no item returned
}
\description{
This function is to organize ismc PSP data
for the sequential compilation.
}
\author{
Yong Luo
}
