% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pspHT.R
\docType{methods}
\name{pspHT}
\alias{pspHT}
\title{Fit height for PSP nonHT data using different method}
\usage{
pspHT(treeData, method = "bestMEM", coeffs)
}
\arguments{
\item{method}{character, Method to derive height, currently it supports
\code{bestHeightModel}.}

\item{coeffs}{data.table, Coefficient table.}

\item{treesData}{data.table, Contains both height trees and nonHT trees in PSPs.}
}
\value{
Full list of trees with derived height.
}
\description{
This function is to fit height for the nonHT trees in PSPs.
}
\author{
Yong Luo
}
