% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSPInit_clusterplot.R
\docType{methods}
\name{PSPInit_clusterplot}
\alias{PSPInit_clusterplot}
\title{Load and select cluster and plot level data for PSP compilation}
\usage{
PSPInit_clusterplot(dataSourcePath)
}
\arguments{
\item{dataSourcePath}{character, Specifies the path that directs to the VRI original data soruce, i.e.,
\code{//Mayhem/GIS_TIB/RDW/RDW_Data2/Work_Areas/VRI_ASCII_PROD/vri_sa}.}
}
\value{
A data table that contains key information at cluster/plot level and compiler log file.
}
\description{
This function prepares the cluster/plot-level inputs for PSP compiler. Specifically, it standardizes names for
the variables; reports and removes the duplicate observations at cluster, cluster/plot.
}
\author{
Yong Luo
}
