% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSPInit_nonHTTree.R
\docType{methods}
\name{PSPInit_nonHTTree}
\alias{PSPInit_nonHTTree}
\title{Load nonHT trees for PSP compilation}
\usage{
PSPInit_nonHTTree(clusterplotHeader, dataSourcePath)
}
\arguments{
\item{clusterplotHeader}{data.table, Cluster and plot level attributes, an output from \code{\link{VRIInit_clusterplot}}.}

\item{dataSourcePath}{character, Specifies the path that directs to the VRI original data soruce, i.e.,
\code{//Mayhem/GIS_TIB/RDW/RDW_Data2/Work_Areas/VRI_ASCII_PROD/vri_sa}.}
}
\value{
A data table that contains auxiliary plot tree data.
}
\description{
This function loads nonHT tree based on cluster/plot header.
}
\author{
Yong Luo
}
