% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pspTreeCheck.R
\docType{methods}
\name{pspTreeCheck}
\alias{pspTreeCheck}
\title{Check psp trees for the abnormal observations}
\usage{
pspTreeCheck(treeData)
}
\arguments{
\item{treesData}{data.table, Contains both height trees and nonHT trees in PSPs.}
}
\value{
Full list of trees with flags.
}
\description{
This function is to check psp trees in terms of dramatic
change of size, live dead code, species and missing observation.
}
\author{
Yong Luo
}
