% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankingMatrix.R
\docType{methods}
\name{rankingMatrix}
\alias{rankingMatrix}
\title{This function is to calculate rating for faib sample sites and to rank them by rating values}
\usage{
rankingMatrix(archivedPSPPath, archivednonPSPPath = NULL, useOldCellKey = TRUE)
}
\arguments{
\item{archivedPSPPath}{character, The path to the compiled PSP data, which is configured
and outputed from \code{ISMCCompiler}.}

\item{archivednonPSPPath}{character, The path to the compiled nonPSP data, which is configured
and outputed from \code{ISMCCompiler}.}

\item{useOldCellKey}{logical, Indicates if using the old cell key from original sas ranking. Default is \code{TRUE}.}
}
\value{
A list of ranking tables including
1) Psp_netdown_summary: net down summary
2) Protected_psp_summary: a summary of protect psp sites by protect code
3)	Remeasured_psp_summary: a summary of remeasured psp sites by last msmt year group
4)	data_source: a description of data used for ranking
5) Ranking_psp: the ranking for psp sample sites
6)	Ranking_matrix: all the raw data and intermediate data used for calculating the rating values
}
\description{
To rank psp sample sites
}
\note{
As the ISMCCompiler could not produce the same cell key as the previously used. The
function allows using a lookup table to populate the cell key to reach consistency.
However, the cell key is further updated using the site status code from ISMC.
}
\author{
Xinjia(Bridget) Guo and Yong Luo
}
