% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readHandHeld.R
\docType{methods}
\name{readHandHeld}
\alias{readHandHeld}
\title{load json data from handheld and save tables into target folder}
\usage{
readHandHeld(
  fileName,
  savePath = getwd(),
  saveFormat = "rdata",
  saveName = "myHandHeldFile",
  overWrite = FALSE
)
}
\arguments{
\item{fileName}{character, JSON file name.}

\item{savePath}{character, A folder path user wants to save all the outputs.
Default is the current work directory.}

\item{saveFormat}{character, Specifies the format user wants to save. Currently, the
function supports \code{xlsx} and \code{rdata}. Default is
\code{rdata}.}

\item{saveName}{character, Specifies the save name.}

\item{overWrite}{logical, Determine if the file with same name as user specifies
will be overwritten. Default is \code{FALSE}.}
}
\value{
No value will be returned. All the outputs will be saved in a target folder.
}
\description{
This function is to load JSON file data from FAIB new handheld.
}
\examples{
\dontrun{
## given, in the current work directory, there is a ismc-file-transfer.json file
   icmcjsonFile <- "ismc-file-transfer.json"
## create a folder in current work directory and save the outputs into it
   dir.create("ismc_test_folder")

   readHandHeld(fileName = icmcjsonFile,
                savePath = file.path(getwd(), "ismc_test_folder"))
}
}
\author{
Yong Luo
}
