% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regRatioDataSelect.R
\docType{methods}
\name{regRatioDataSelect}
\alias{regRatioDataSelect}
\title{Select the data for regression and ratio}
\usage{
regRatioDataSelect(sampledata, alltreedata, usage)
}
\arguments{
\item{sampledata}{data.table, The data contains the sample level information.}

\item{alltreedata}{data.table, The is the data pool which contains the full/enhanced trees
and H-enhanced trees.}

\item{usage}{character, Specifies compiler name. It supports \code{ismc} and \code{vgis}.}
}
\value{
Selected data for regression and ratio.
}
\description{
This function selects the data for fitting regression model between basal area and
whole stem volume, and for ratios to whole stem volume in VRI compiler. The regreesion
method is used for derive whole stem volume for the trees that just have DBH information.
The ratio method is used to derive netted merchantable volume for the trees that
do not have call grading information. The data selection should be done annually at the end of
every March to allow newer and higher quaulity data enter the data.
}
\author{
Yong Luo
}
