% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskGroupDeriver.R
\docType{methods}
\name{riskGroupDeriver}
\alias{riskGroupDeriver}
\alias{riskGroupDeriver,character,character,character,numeric,character-method}
\alias{riskGroupDeriver,character,character,missing,missing,character-method}
\title{Derive risk group for standard sample compilation/data}
\usage{
riskGroupDeriver(species, pathIndex, series, height, method)

\S4method{riskGroupDeriver}{character,character,character,numeric,character}(species, pathIndex, series, height, method)

\S4method{riskGroupDeriver}{character,character,missing,missing,character}(species, pathIndex, series, height, method)
}
\arguments{
\item{species}{character, Tree basic species code, which is SP0 in VRI original data.}

\item{pathIndex}{character, A character with length of 8, consists of 0 or 1.}

\item{series}{character, DWB series. It is a lenght of 2 number character and can be derived using
\code{\link{getDWBSeries}} function.}

\item{height}{numeric, Total tree height.}

\item{method}{character, Specifies the method between \code{FIZ} and \code{KBEC} to categorize the risk group. The \code{FIZ} method derives risk
group by \code{species}, \code{pathIndex}, \code{series} and \code{height}.
\code{KBEC} method derives the risk group using \code{species} and \code{pathIndex}.}
}
\value{
Risk group, which is character
}
\description{
will refine. This function is equivalent to risk_grp.sas macro.
}
\author{
Yong Luo
}
