% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplePlotCompilation.R
\docType{methods}
\name{samplePlotCompilation}
\alias{samplePlotCompilation}
\title{Compile sample and plot level information}
\usage{
samplePlotCompilation(compilationType, dataSourcePath, mapPath, coeffPath)
}
\arguments{
\item{compilationType}{character, either \code{PSP} or \code{nonPSP}. If it is \code{PSP}, it
is consistent with original PSP compiler, otherwise, it
is consistent with VRI compiler.}

\item{dataSourcePath}{character, Specifies the path that contains prepared data from raw data.}

\item{mapPath}{character, Specifies the path dependent maps are stored.}

\item{coeffPath}{character, Specifies the path dependent coeffs are stored.}
}
\value{
A data table that contains key information at cluster/plot level and compiler log file.
}
\description{
This function is to compile sample and plot information.
}
\author{
Yong Luo
}
