% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siteAgeCompiler.R
\docType{methods}
\name{siteAgeCompiler}
\alias{siteAgeCompiler}
\alias{siteAgeCompiler,data.table,character-method}
\title{Compile breast age, total age, and site index where possible-VRI specific}
\usage{
siteAgeCompiler(siteAgeData, compilationType)

\S4method{siteAgeCompiler}{data.table,character}(siteAgeData, compilationType)
}
\arguments{
\item{siteAgeData}{data.table, Site age data with plot header information.
An output from \code{\link{VRIInit_siteTree}} function.}

\item{compilationType}{character, Compilation type.}
}
\value{
A data table and a log file.
}
\description{
This function takes site age tree data ie., \code{vi_h}, an output of \code{\link{VRIInit_siteTree}} to
compute the breast height age, total age, and site index where possible. This function is
equivalent to site_age.sas. The function heavily depends on site tools program.
}
\author{
Yong Luo
}
