% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siteAgeCorrection.R
\docType{methods}
\name{siteAgeCorrection}
\alias{siteAgeCorrection}
\title{Bored age correction for multiple measurments based on measurement interval}
\usage{
siteAgeCorrection(vih)
}
\arguments{
\item{vih}{data.table, The table contains the site trees in the database.}
}
\value{
A site tree table that is equivalent to original vih,
but with bored age corrected
}
\description{
Correct the bored age if the difference of ages between two measurements does
not match the years of measurement interval
}
\note{
\enumerate{
\item for the site trees with the age measurement code of PRE, the current age is corrected
based on the previous bored age.
2) for the site trees with multiple drills, the age of last measurement is used as a
reference point, and the previous ages are adjusted.
3) some ages of site trees are manuallly corrected due to complexity of measurements (see comments in the codes)
}
}
\author{
Yong Luo
}
