% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siteAgeSummary.R
\docType{methods}
\name{siteAgeSummary}
\alias{siteAgeSummary}
\alias{siteAgeSummary,data.table-method}
\title{Summarize site age data by cluster and cluster/species-VRI specific}
\usage{
siteAgeSummary(cpldSiteAgeData)

\S4method{siteAgeSummary}{data.table}(cpldSiteAgeData)
}
\arguments{
\item{cpldSiteAgeData}{data.table, Compiled site age tree data, an output of \code{\link{siteAgeCompiler}}.}
}
\value{
Two data tables: cl_ah is the age/height summary at cluster level and spc_ah is the age/height summary
at cluster and species level
}
\description{
This function takes compiled site age tree data, an output of \code{\link{siteAgeCompiler}}, to
derive mean age and height results. The compiled data must have breast height age,
total age, and site index. This function is equivalent to \code{mean_htl.sas}.
}
\author{
Yong Luo
}
