% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siteAgeSummary_PSP.R
\docType{methods}
\name{siteAgeSummary_PSP}
\alias{siteAgeSummary_PSP}
\title{Summarize site age data by cluster and cluster/species for PSP data}
\usage{
siteAgeSummary_PSP(cpldSiteAgeData, treemsmt)
}
\arguments{
\item{cpldSiteAgeData}{data.table, Compiled site age tree data, an output of \code{\link{siteAgeCompiler}}.}

\item{treemsmt}{data.table, Tree measurement data.}
}
\value{
Two data tables: cl_ah is the age/height summary at cluster level and spc_ah is the age/height summary
at cluster and species level
}
\description{
This function takes compiled site age tree data, an output of \code{\link{siteAgeCompiler}}, to
derive mean age and height results. The compiled data must have breast height age,
total age, and site index.
}
\author{
Yong Luo
}
