% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smallTreeVolSmry.R
\docType{methods}
\name{smallTreeVolSmry}
\alias{smallTreeVolSmry}
\alias{smallTreeVolSmry,data.table,data.table-method}
\title{Summarizes the volume for small trees - VRI specific}
\usage{
smallTreeVolSmry(smallTreeData, smallTreePlotHeader)

\S4method{smallTreeVolSmry}{data.table,data.table}(smallTreeData,
  smallTreePlotHeader)
}
\arguments{
\item{smallTreeData}{data.table, Small tree data. This data is from card f, i.e., \code{vi_f}.}

\item{smallTreePlotHeader}{data.table, Plot header data for stump and small tree data. The data is from card e, i.e.,
\code{vi_e}.}
}
\value{
Two tables: stmp_c is summarized volume at cluster level; stmp_cs is summarized volume at cluster/species level.
}
\description{
Calculates and summarizes volume of small trees at both cluster/species level and cluster.
This function is equivalent to \code{sml_tree.sas} in original compiler.
}
\author{
Yong Luo
}
