% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spAttriFromRegCompt.R
\name{spAttriFromRegCompt}
\alias{spAttriFromRegCompt}
\title{get tsa, bec, fiz and ownership attributes based on region and compartment map}
\usage{
spAttriFromRegCompt(regionCompartMap, spAttributeMap, mapName)
}
\arguments{
\item{regionCompartMap}{sf, regioncompartment map.}

\item{spAttributeMap}{sf, spatial attribute map.}

\item{mapName}{character, spatial attribute map name, must be one of \code{BEC}, \code{TSA},
\code{FIZ}, \code{TFL}, and \code{Ownership}.}
}
\value{
table of spatial attributes corresponding to region and compartment.
}
\description{
get tsa, bec, fiz and ownership attributes based on region and compartment map
}
\note{
The spatial attributes must be one-on-one match to a unique combination of region
and compartment
}
\author{
Yong Luo
}
