% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spCorr.R
\docType{methods}
\name{spCorr}
\alias{spCorr}
\title{tree level species correct for ISMC database, based on bec and bec subzone}
\usage{
spCorr(BECInfor, dataSourcePath)
}
\arguments{
\item{BECInfor}{data.table, Cluster and plot level attributes, an output from \code{\link{VRIInit_clusterplot}}.}

\item{dataSourcePath}{character, Specifies the path that directs to compilation_sa.}
}
\value{
A data table that contains auxiliary plot tree data.
}
\description{
This function is to correct species.
}
\author{
Yong Luo
}
