% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesComp_byC.R
\docType{methods}
\name{speciesComp_byC}
\alias{speciesComp_byC}
\alias{speciesComp_byC,data.table,character,numeric,logical-method}
\alias{speciesComp_byC,data.table,character,numeric,missing-method}
\title{Calculates species composition based on cluster/species summary}
\usage{
speciesComp_byC(CSSmryTable, basedOn, speciesMaxNO, smallTreeCompile)

\S4method{speciesComp_byC}{data.table,character,numeric,logical}(CSSmryTable, basedOn, speciesMaxNO, smallTreeCompile)

\S4method{speciesComp_byC}{data.table,character,numeric,missing}(CSSmryTable, basedOn, speciesMaxNO)
}
\arguments{
\item{CSSmryTable}{data.table, Summarized volume components for both measured and counted trees at cluster and species level.
See \code{\link{volSmry_byCS}} for details.}

\item{basedOn}{character, Specifies which component is used for species composition summary.}

\item{speciesMaxNO}{numeric, Maximum number of species entries to calculate.}

\item{smallTreeCompile}{logical, Defines whether the function is used for calculate species composition
for small trees. If missing, \code{FALSE} is used.}
}
\value{
A data table that contains species composition at cluster level.
}
\description{
Calculates species composition at cluster level based on cluster/species summary.
The cluster/species-level summaries is an output of \code{\link{volSmry_byCS}}
function. This function is equivalent to \code{sp_comp.sas} in original compiler.
}
\author{
Yong Luo
}
