% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesCorrection.R
\docType{methods}
\name{speciesCorrection}
\alias{speciesCorrection}
\title{Species correction, based on Rene and Dan's correction routine. See emails on
May 12th, 2021.}
\usage{
speciesCorrection(species, BEC, BEC_subzone)
}
\arguments{
\item{species}{character, Specifies the original species.}

\item{BEC}{character, BEC zone.}

\item{BEC_subzone}{character, BEC subzone.}
}
\value{
corrected species code
}
\description{
This function is to correct species codes
}
\author{
Yong Luo
}
