% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stumpVolSmry.R
\docType{methods}
\name{stumpVolSmry}
\alias{stumpVolSmry}
\alias{stumpVolSmry,data.table,data.table-method}
\title{Summarizes the volume for stumps - VRI specific}
\usage{
stumpVolSmry(stumpData, stumpPlotHeader)

\S4method{stumpVolSmry}{data.table,data.table}(stumpData, stumpPlotHeader)
}
\arguments{
\item{stumpData}{data.table, Stump data. This data is from card g, i.e., \code{vi_g}.}

\item{stumpPlotHeader}{data.table, Plot header data for stump and small tree data. The data is from card e, i.e.,
\code{vi_e}.}
}
\value{
Two tables: stmp_c is summarized volume at cluster level; stmp_cs is summarized volume at cluster/species level.
}
\description{
Calculates stump volume at cluster/species level and cluster.
This function is equivalent to \code{stmpvol.sas} in original compiler.
}
\author{
Yong Luo
}
