% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toWSVRatio.R
\docType{methods}
\name{toWSVRatio}
\alias{toWSVRatio}
\title{To toWSV ratio for each of strata}
\usage{
toWSVRatio(inputData, needCombs, minDBH = 10, minObs = 30)
}
\arguments{
\item{inputData}{data.table, The data for deriving ratios, that must contain full and enhanced
trees.}

\item{needCombs}{data.table, The combinations of BEC+SP0+LV_D are needed to derive coefficients.}

\item{minDBH}{numeric, The minimum DBH for selecting trees to derive ratios. If missing,
10 cm will be used.}

\item{minObs}{numeric, The minimum samples size for a stratum. If missing,
30 observations will be used.}
}
\value{
ratio table
}
\description{
This function takes the selected data for derive ratio for each of strata by
BEC+SP0+LV_D
}
\author{
Yong Luo
}
