% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemsmtEditing.R
\docType{methods}
\name{treemsmtEditing}
\alias{treemsmtEditing}
\title{Edit tree measurements for a repeatedly measured data.}
\usage{
treemsmtEditing(treemsmts, sitevisits)
}
\arguments{
\item{treemsmts}{data.table, Tree measurements. The table must contain unique tree id, i.e.,
\code{SITE_IDENTIFIER}, \code{PLOT}, and \code{TREE_NUMBER}, and
\code{VISIT_NUMBER}.
And key measurements including live/dead (\code(TREE_EXTANT_CODE)),
diameter information (\code{DIAMETER} and \code{DIAMETER_MEASMT_HEIGHT}),
\code{TREE_CLASS_CODE}, \code{BROKEN_TOP_IND}, \code{CROWN_CLASS_CODE},
and \code{TREE_SPECIES_CODE}.}

\item{sitevisits}{data.table, Site visit information including \code{SITE_IDENTIFIER} and \code{VISIT_NUMBER}.}
}
\value{
return a table after editing.
}
\description{
This function takes tree-level measurements and edits live/dead codes,
diameter, species, and add tree measurements if the measurements are missing.
}
\note{
\enumerate{
\item For missing live/dead code, using the next live/dead code if it is present. Otherwise,
assign dead (D).
2. In the case a tree is observed dead at a visit and change to live in the next visit,
the dead code (D) will be changed to live (L).
3. If there is a missing measurement in between two visits. A measurement will be added
with diameter assigned as mean of previous and next diameter and live/dead will be populated using
next measurement.
4. If there is a missing measurement at tail, i.e., a tree was seen live for a given visit but
without reaching the last visit for a given site. A measurement for next visit will be added with
diameter populated using previous diameter and live/dead code will be assigned as D.
5. If species code changes during visits, the species code at last visit will be used to correct
species code throughout all visits.
6. If at a tree's last visit, there is no diameter information and it is marked as live tree.
The live/dead code will be changed to D.
7. If a tree is dead at a given visit and without diameter, and a tree was live at previous
visit with diameter, the diameter at previous visit will be used to populate diameter at this
visit.
8. If a tree is broken top tree for a given visit, this tree must be broken top tree since
then.
9. If crown_class_code is missing at a visit and it was present at previous visit,
the previous crown_class_code will be used to populate for this visit.
}
}
\author{
Yong Luo
}
