% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeVolEst_RegRatio_new.R
\docType{methods}
\name{treeVolEst_RegRatio_new}
\alias{treeVolEst_RegRatio_new}
\title{Estimate volume for H-enhanced and non-enhanced trees-VRI specific}
\usage{
treeVolEst_RegRatio_new(nonVolTrees, BA_WSVCoeff, ratioTable)
}
\arguments{
\item{nonVolTrees}{data.table, H-enhanced trees and non-enhanced trees.}

\item{BA_WSVCoeff}{list, Contains fixed and random coefficients the WSV-BA equations by project group \code{PRJ_GRP}, live and dead status \code{LV_D},
stand and falling status \code{SF_COMPILE} and species code \code{SP0}. The table can be
generated using \code{\link{WSV_BARegression}}.}

\item{ratioTable}{list, Specifies \code{toWSV} ratio by project group \code{PRJ_GRP}, live and dead status \code{LV_D},
stand and falling status \code{SF_COMPILE} and species code \code{SP0}.
The table can be generated using \code{\link{toWSVRatio}}.}
}
\value{
A data table that has compiled non volume trees.
}
\description{
This function estimates the volumes for JH-enhanced and non-enhanced trees using \code{BA-WSV} equation and
\code{toWSV} ratio methods. For H-enhanced trees, the whole stem volume and gross merchantable volume are already calculated directly using
taper equations; and rest of volume components will be calculated using ratio method in this function.
For non-enhanced trees, the whole stem volume is derived using regression equation between basal area
and whole stem volume and the rest of volume components will be computed using ratio method in this function.
}
\author{
Yong Luo
}
