% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateSA_vegcomp.R
\docType{methods}
\name{updateSA_vegcomp}
\alias{updateSA_vegcomp}
\title{Update projected stand age from veg comp rank 1 layer}
\usage{
updateSA_vegcomp(
  compilationType,
  coeffPath,
  bcgwUserName,
  bcgwPassword,
  sampleSites,
  sampleMsmts
)
}
\arguments{
\item{compilationType}{character, either \code{PSP} or \code{nonPSP}. If it is \code{PSP}, it
is consistent with original PSP compiler, otherwise, it
is consistent with VRI compiler.}

\item{coeffPath}{character, Specifies the path dependent coeffs are stored.}

\item{bcgwUserName}{character, User name to access to bcgw database.}

\item{bcgwPassword}{character, Password to access to bcgw database.}

\item{sampleSites}{data.table, sample sites A data table must contain site_identifier, bc albers coordinates.}

\item{sampleMsmts}{data.table, sample measurements A data table must contain site_identifier, bc albers coordinates.}
}
\value{
A data table with feature_id, proj_age_1 and projected_date.
}
\description{
This function is to update projected stand age from vegcomp rank 1 layer.
}
\author{
Yong Luo
}
