% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateSpatial.R
\name{updateSpatial}
\alias{updateSpatial}
\title{Used for updating spatial attributes based on locations}
\usage{
updateSpatial(samplesites, mapPath)
}
\arguments{
\item{samplesites}{data.table A table that contains spatial loctions. Currently it is
designed for accepting UTM coordinates. Therefore, zone, northing
and easting must be provided.}

\item{mapPath}{character The path to fiz map.}
}
\value{
a table that contains spatial attributes of bc albers, longitude/latitude,
TSA, BEC, FIZ, TFL and OWNERSHIP.
}
\description{
Used for updating spatial attributes based on locations.
}
\author{
Yong Luo
}
