% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateSpatial_badUTM_PSP.R
\name{updateSpatial_badUTM_PSP}
\alias{updateSpatial_badUTM_PSP}
\title{Used for updating spatial attributes for the sites with bad utm}
\usage{
updateSpatial_badUTM_PSP(mapPath, samplesites)
}
\arguments{
\item{mapPath}{character The path to fiz map.}

\item{samplesites}{data.table A table that contains spatial loctions. Currently it is
designed for accepting UTM coordinates. Therefore, zone, northing
and easting must be provided.}

\item{compilationType}{character, either \code{PSP} or \code{nonPSP}. If it is \code{PSP}, it
is consistent with original PSP compiler, otherwise, it
is consistent with VRI compiler.}
}
\value{
a table that contains spatial attributes of bc albers, longitude/latitude,
TSA, BEC, FIZ, TFL and OWNERSHIP.
}
\description{
Used for updating spatial attributes for the sites with bad utm.
}
\note{
This routine is based on discussion with Dan and Anya on 2023-10-05
}
\author{
Yong Luo
}
