% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vicPrep.R
\docType{methods}
\name{vicPrep}
\alias{vicPrep}
\title{prepare vi_c for compilation}
\usage{
vicPrep(compilationType, clusterplotHeader, dataSourcePath, walkThru = TRUE)
}
\arguments{
\item{compilationType}{character, either \code{PSP} or \code{nonPSP}. If it is \code{PSP}, it
is consistent with original PSP compiler, otherwise, it
is consistent with VRI compiler.}

\item{clusterplotHeader}{data.table, Cluster and plot-level attributes.}

\item{dataSourcePath}{character, Specifies the path that directs to the compilation_sa.}

\item{walkThru}{logical, Indicates whether walkthrough sampling protocal is used,
Tree weight is determined by walkthrough method. In walkthrough
method, a tree is identified as \code{NA} (no walkthrough applied),
\code{O} for out tree (not counted), and \code{W} for double counted tree.}
}
\value{
A data table that contains tree-level information. A log file that describes the detailed process.
}
\description{
This function selects the tree-level data from vi_c (cardc) based on selected cluster/plot headers.
Additonally, the function calculates basal area and tree per ha factor.
}
\author{
Yong Luo
}
