% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vidPrep.R
\docType{methods}
\name{vidPrep}
\alias{vidPrep}
\title{Load and select trees that have loss factor information-VRI specific}
\usage{
vidPrep(dataSourcePath)
}
\arguments{
\item{dataSourcePath}{character, Specifies the path that directs to the VRI original data soruce, i.e.,
\code{//Mayhem/GIS_TIB/RDW/RDW_Data2/Work_Areas/VRI_ASCII_PROD/vri_sa}.}
}
\value{
A data table that contains loss factor data. A log file documents the detailed process
}
\description{
This function loads and selects trees that have loss factor information (\code{vi_d}, cardd) based on
selected trees from vi_c.
}
\author{
Yong Luo
}
