% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vihPrep.R
\docType{methods}
\name{vihPrep}
\alias{vihPrep}
\title{to populate bored age for site trees using last measurements, set meas_cod = "FROM_LAST
and to correct bored age using last measurements, set meas_cod = *-ADJ_FROM_LAST}
\usage{
vihPrep(msmtInterval, siteAgeTrees)
}
\arguments{
\item{msmtInterval}{data.table, contains sample visits.}

\item{siteAgeTrees}{data.table, sample trees.}
}
\value{
A data table
}
\description{
This function is to populate bored age for site trees using last measurements, set meas_cod = "FROM_LAST
and to correct bored age using last measurements, set meas_cod = *-ADJ_FROM_LAST
}
\author{
Yong Luo
}
