% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viiPrep.R
\docType{methods}
\name{viiPrep}
\alias{viiPrep}
\title{Load and select auxiliary plot trees-VRI specific}
\usage{
viiPrep(compilationType, clusterplotHeader, dataSourcePath)
}
\arguments{
\item{compilationType}{character, either \code{PSP} or \code{nonPSP}. If it is \code{PSP}, it
is consistent with original PSP compiler, otherwise, it
is consistent with VRI compiler.}

\item{clusterplotHeader}{data.table, Cluster and plot level attributes, an output from \code{\link{VRIInit_clusterplot}}.}

\item{dataSourcePath}{character, Specifies the path that directs to the VRI original data soruce.}
}
\value{
A data table that contains auxiliary plot tree data.
}
\description{
This function loads and selects auxiliary data (\code{vi_i}, cardi) based on cluster/plot header.
}
\author{
Yong Luo
}
