% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volSmry_byC.R
\docType{methods}
\name{volSmry_byC}
\alias{volSmry_byC}
\alias{volSmry_byC,data.table-method}
\title{Summarize volume components at cluster level-VRI specific}
\usage{
volSmry_byC(volSmryByCS)

\S4method{volSmry_byC}{data.table}(volSmryByCS)
}
\arguments{
\item{volSmryByCS}{data.table, Summarized volume components for both measured and counted trees.
See \code{\link{volSmry_byCS}} for details.}
}
\value{
A data table
}
\description{
This function summarizes the cluster-level volume components using cluster/species-level
summaries. The cluster/species-level summaries is an output of \code{\link{volSmry_byCS}}
function.
}
\author{
Yong Luo
}
