% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volSmry_byCS.R
\docType{methods}
\name{volSmry_byCS}
\alias{volSmry_byCS}
\alias{volSmry_byCS,data.table,numeric,character,character-method}
\alias{volSmry_byCS,data.table,missing,character,character-method}
\alias{volSmry_byCS,data.table,numeric,missing,character-method}
\alias{volSmry_byCS,data.table,numeric,numeric,character-method}
\alias{volSmry_byCS,data.table,numeric,character,missing-method}
\alias{volSmry_byCS,data.table,missing,missing,missing-method}
\title{Summarize volume components per hectare by cluster and species-VRI specific}
\usage{
volSmry_byCS(treeMC, utilLevel, weirdUtil, equation)

\S4method{volSmry_byCS}{data.table,numeric,character,character}(treeMC,
  utilLevel, weirdUtil, equation)

\S4method{volSmry_byCS}{data.table,missing,character,character}(treeMC,
  weirdUtil, equation)

\S4method{volSmry_byCS}{data.table,numeric,missing,character}(treeMC,
  utilLevel, equation)

\S4method{volSmry_byCS}{data.table,numeric,numeric,character}(treeMC,
  utilLevel, weirdUtil, equation)

\S4method{volSmry_byCS}{data.table,numeric,character,missing}(treeMC,
  utilLevel, weirdUtil)

\S4method{volSmry_byCS}{data.table,missing,missing,missing}(treeMC)
}
\arguments{
\item{treeMC}{data.table, Tree-level compiled data for all volume trees.}

\item{utilLevel}{numeric, Utilization levels. Default is 4.}

\item{weirdUtil}{character, Weird util. Default is \code{No}. Otherwise need to be specified as a number.}

\item{equation}{character, Specifies whether the compiler is based on \code{KBEC} or \code{KFIZ}.
Default is \code{KBEC}.}
}
\value{
A data table summarizes volume components by cluster and species. Equevalent to \code{smy_cs}.
}
\description{
Summarizes volume components per hectare by cluster and species.
The function is last part of \code{vol_ha_2017.sas}.
}
\author{
Yong Luo
}
