% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VRICompiler.R
\docType{methods}
\name{VRICompiler}
\alias{VRICompiler}
\title{VRI compiler - VRI specific}
\usage{
VRICompiler(
  oracleUserName,
  oraclePassword,
  asciiTxtPath = "//albers/gis_tib/VRI/RDW/RDW_Data2/Work_Areas/VRI_ASCII_PROD/vri_raw",
  compilationPath =
    "//albers/gis_tib/VRI/RDW/RDW_Data2/Work_Areas/VRI_ASCII_PROD/FromRCompiler",
  coeffPath =
    "//albers/gis_tib/VRI/RDW/RDW_Data2/Work_Areas/VRI_ASCII_PROD/FromRCompiler/Coeffs",
  fizmapPath = "//spatialfiles2.bcgov/work/for/vic/hts/dam/workarea/data/infrastructure",
  fizmapName = "FIZ_REG_COMPARTMENT",
  fizmapFormat = "gdb",
  equation = "KBEC",
  walkThru = TRUE,
  logMinLength = 0.1,
  stumpHeight = 0.3,
  breastHeight = 1.3,
  UTOPDIB = 10,
  utilLevel = 4,
  weirdUtil = "No"
)
}
\arguments{
\item{oracleUserName}{character, User name to access to oracle database.}

\item{oraclePassword}{character, Password to access to oracle database.}

\item{asciiTxtPath}{character, Path to ascii txt files. By default, the arguement is specified to
\code{//albers/gis_tib/VRI/RDW/RDW_Data2/Work_Areas/VRI_ASCII_PROD/vri_raw}.
However, user can modify.}

\item{compilationPath}{character, Specifies the path that stores all the data/processes. By specifying this,
four folders will be created to record all the data/processes. Specifically,
raw_from_oracle stores the data just after oracle and ascii without editing;
compilation_sa stores key data (not all) that after editing and before volume and age compilation;
compilation_db stores compiled results for volume and age compilation at both tree level
and cluater level;
Archive_YYYYMMDD achives all the data mentioned above for the future use or reference.
By default, this path is set as \code{//albers/gis_tib/VRI/RDW/RDW_Data2/Work_Areas/VRI_ASCII_PROD/RCompilation},
which is consistent with our rdw system.}

\item{coeffPath}{character, Specifies the path that stores coefficients and spatial lookup tables.}

\item{fizmapPath}{character, Specifies the path to forest inventory zone map. By default,
it is set to \code{//spatialfiles2.bcgov/work/for/vic/hts/dam/workarea/data/infrastructure},
which is maintained by FAIB employee.}

\item{fizmapName}{character, Specifies the name of forest inventory zone map. By default,
it is set to \code{FIZ_REG_COMPARTMENT}, which is maintained by FAIB employee.}

\item{fizmapFormat}{character, Specifies the format of forest inventory zone map. Currently, it can be specified
as \code{gdb} for geodatabase format and \code{shp} for shapefile format.
By default, it is set to \code{gdb}, which is maintained by FAIB employee.}

\item{equation}{character, Specifies the taper equation that is used for compiler. Currently supports
BEC-based (\code{KBEC}) and FIZ-based (\code{KFIZ}).}

\item{walkThru}{logical, Speciefies whether the data had been collected using work through method. Default is \code{TRUE},
if it is not specified.}

\item{logMinLength}{numeric, Specifies minimum length of log when doing log length adjustment,
see \code{\link{logMatrixAdjustment}} for details. If missing 0.1 is used.}

\item{stumpHeight}{numeric, Stump height. If missing 0.3 is used.}

\item{breastHeight}{numeric, Breast height. If missing 1.3 is used.}

\item{UTOPDIB}{numeric, Threshold inside-bark diameter for merchantable volume. If missing, UTOPDIB is 10.}

\item{utilLevel}{numeric, Specifies utilization level in summrizing tree volumes at cluster and species level. Default is 4.}

\item{weirdUtil}{character, Specifies weird utilization in summarizing tree volumes at cluster and species level.
Default is \code{no}, if missing. Otherwise, a number should be provided.}
}
\value{
This function compiles data and save outputs in \code{compilationPaths$compilation_db} and no file is returned.
}
\description{
This function compiles VRI data by calling specific VRI functions. Unlike the original
compiler (i.e., SAS compiler), the R version compiler hardcodes all the lookup tables in
the compilation process. Please refer the descriptions for lookup table to see whether
they are same as the original lookup table.
}
\note{
Improvements include:
\enumerate{
\item specifies trees in B plots as height enhanced trees
\item removes the sas-dependent sindex functions
\item introduce the SIndexR package
}
Currently, the compiler supports compilation for the below sample types:
\itemize{
\item{\code{Q: }} {Regular VRI sample with five point clusters design}
\item{\code{T: }} {??, sample protocol and sample design are same as Q samples}
\item{\code{B: }} {??, same plot layout as Q sample, with height is measured for all trees and no call grading information}
\item{\code{M: }} {CMI sample, with all trees have call grading information in the field}
\item{\code{L: }} {LiDAR project, same plot layout and same design but without call grading information}
\item{\code{Y: }} {YSM plots, population between 15 and 50 years}
\item{\code{F: }} {NFI plots, trees measured all DBH, height and call grading}
\item{\code{N: }} {NVAF plots}
\item{\code{A: }} {VRI audit plots}
}
}
\references{
VRI compiler manual
}
\author{
Yong Luo
}
