% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VRIInit_measuredTree.R
\docType{methods}
\name{VRIInit_measuredTree}
\alias{VRIInit_measuredTree}
\title{Load and select fully measured tree data-VRI specific}
\usage{
VRIInit_measuredTree(clusterplotHeader, dataSourcePath, walkThru = TRUE)
}
\arguments{
\item{clusterplotHeader}{data.table, Cluster and plot-level attributes, an output from \code{\link{VRIInit_clusterplot}}.}

\item{dataSourcePath}{character, Specifies the path that directs to the VRI original data soruce, i.e.,
\code{//Mayhem/GIS_TIB/RDW/RDW_Data2/Work_Areas/VRI_ASCII_PROD/vri_sa}.}

\item{walkThru}{logical, Indicates whether walkthrough sampling protocal is used,
Tree weight is determined by walkthrough method. In walkthrough
method, a tree is identified as \code{NA} (no walkthrough applied),
\code{O} for out tree (not counted), and \code{W} for double counted tree.}
}
\value{
A data table that contains tree-level information. A log file that describes the detailed process.
}
\description{
This function selects the tree-level data from vi_c (cardc) based on selected cluster/plot headers.
Additonally, the function calculates basal area and tree per ha factor.
}
\author{
Yong Luo
}
