% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VRIInit_siteTree.R
\docType{methods}
\name{VRIInit_siteTree}
\alias{VRIInit_siteTree}
\title{Load and select site trees-VRI specific}
\usage{
VRIInit_siteTree(clusterplotHeader, dataSourcePath)
}
\arguments{
\item{clusterplotHeader}{data.table, contains cluster/plot-level attributes, an output from \code{\link{VRIInit_clusterplot}}.}

\item{dataSourcePath}{character, Specifies the path that directs to the VRI original data soruce, i.e.,
\code{//Mayhem/GIS_TIB/RDW/RDW_Data2/Work_Areas/VRI_ASCII_PROD/vri_sa}.}
}
\value{
A data table that contains site tree data information. A log file documents the detailed process
}
\description{
This function connects site tree data (vi_h, cardh) to selected cluster/plot-level data.
Site tree data is located in  \code{//Mayhem/GIS_TIB/RDW/RDW_Data2/Work_Areas/VRI_ASCII_PROD/vri_sa}
}
\note{
VRI specific
}
\author{
Yong Luo
}
