% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VRISummaries.R
\docType{methods}
\name{VRISummaries}
\alias{VRISummaries}
\alias{VRISummaries,data.table,data.table,numeric,character,character-method}
\title{Summarize the tree-level data at cluster or cluster/species level-VRI specific}
\usage{
VRISummaries(allVolumeTrees, clusterPlotHeader, utilLevel, weirdUtil,
  equation)


  \S4method{VRISummaries}{data.table,data.table,numeric,character,character}(allVolumeTrees,
  clusterPlotHeader, utilLevel, weirdUtil, equation)
}
\arguments{
\item{allVolumeTrees}{data.table, All tree data from vi_c and vi_i that have been compiled with
tree volume.}

\item{clusterPlotHeader}{data.table, Cluster and plot-level information. An output of \code{\link{VRIInit_clusterplot}}.}

\item{utilLevel}{numeric, Utilization levels.}

\item{weirdUtil}{character, Weird util. Default is No. Otherwise need to be specified as a number.}

\item{equation}{character, Specifies whether the compiler is based on \code{KBEC} or \code{KFIZ}.}
}
\value{
Cluster and species-level volume summaries; cluster-level volume summaries; cluster-level height summaries;
cluster-level species composition summaries and log file.
}
\description{
Summarizes the compiled tree data (including both enhanced tree data and non-enhanced tree data) at
cluster level. This function is equevalent to the summary part in sas compiler in \code{cp_vegi_2017.sas}.
Different from the original compiler, this function outputs the summaries by summarized components,
rather than putting all together.
}
\author{
Yong Luo
}
