% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBH_Height_MEM.R
\name{DBH_Height_MEM}
\alias{DBH_Height_MEM}
\title{Develop the mixed effect models between DBH and height by strata.}
\usage{
DBH_Height_MEM(userName, passWord, env, mapPath, coeffSavePath)
}
\arguments{
\item{userName}{character, Specifies a valid user name in ISMC Oracle database.}

\item{passWord}{character, Specifies the password to the user name.}

\item{env}{character, Specifies which environment the data reside. Currently,
the function supports \code{INT} (intergration)
and \code {TEST} (test) environment.}

\item{coeffSavePath}{character, Specifies the path to save your outputs.}
}
\value{
no item returned
}
\description{
Develop the mixed effect model, and select the best models per strata.
}
\author{
Yong Luo
}
