% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadASCII.R
\docType{methods}
\name{loadASCII}
\alias{loadASCII}
\title{Load the data from ASCII files}
\usage{
loadASCII(txtLocation, saveThem = FALSE, savePath)
}
\arguments{
\item{txtLocation}{character, Specifies the location of ASCII files.}

\item{saveThem}{logical, Specifies whether the loaded data should be saved or returned.
The default value is FALSE, which means the function will not save files
for you.}

\item{savePath}{character, Specifies the path that directs to the VRI original data soruce, i.e.,
\code{//Mayhem/GIS_TIB/RDW/RDW_Data2/Work_Areas/VRI_ASCII_PROD/vri_sa}.}
}
\value{
no files
}
\description{
This function is to read the data from ASCII files (in txt format).
}
\examples{
\dontrun{
loadASCII(txtLocation = "Q:/RDW_Data2/Work_Areas/VRI_ASCII_PROD/vri_raw/to_be_added_next_compilation",
          saveThem = TRUE,
          savePath = "D:/vritest/forascii")

}
}
\author{
Yong Luo
}
