% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateSpatial.R
\name{updateSpatial}
\alias{updateSpatial}
\title{Used for updating spatial attributes based on locations}
\usage{
updateSpatial(
  samplesites,
  fizmapPath,
  fizmapName,
  fizmapFormat,
  tflmapPath,
  tflmapName,
  tflmapFormat,
  ownershipmapPath,
  ownershipmapName,
  ownershipmapFormat
)
}
\arguments{
\item{samplesites}{data.table A table that contains spatial loctions. Currently it is
designed for accepting UTM coordinates. Therefore, zone, northing
and easting must be provided.}

\item{fizmapPath}{character The path to fiz map.}

\item{fizmapName}{character The name of fiz map.}

\item{fizmapFormat}{character The format of fiz map.}

\item{tflmapPath}{character The path to tfl map.}

\item{tflmapName}{character The name of tfl map.}

\item{tflmapFormat}{character The format of tfl map.}

\item{ownershipmapPath}{character The path to ownership map.}

\item{ownershipmapName}{character The name of ownership map.}

\item{ownershipmapFormat}{character The format of ownership map.}
}
\value{
a table that contains spatial attributes of bc albers, longitude/latitude,
TSA, BEC, FIZ, TFL and OWNERSHIP.
}
\description{
Used for updating spatial attributes based on locations.
}
\author{
Yong Luo
}
