% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateSpatial.R
\name{updateSpatial}
\alias{updateSpatial}
\title{Used for updating spatial attributes based on locations}
\usage{
updateSpatial(compilationType, samplesites, mapPath)
}
\arguments{
\item{compilationType}{character, either \code{PSP} or \code{nonPSP}. If it is \code{PSP}, it
is consistent with original PSP compiler, otherwise, it
is consistent with VRI compiler.}

\item{samplesites}{data.table A table that contains spatial loctions. Currently it is
designed for accepting UTM coordinates. Therefore, zone, northing
and easting must be provided.}

\item{mapPath}{character The path to fiz map.}
}
\value{
a table that contains spatial attributes of bc albers, longitude/latitude,
TSA, BEC, FIZ, TFL and OWNERSHIP.
}
\description{
Used for updating spatial attributes based on locations.
}
\author{
Yong Luo
}
